Refactoring scene-integration into separate test files.
